requirements.txt
# Core
python-dateutil>=2.8.2
pytz>=2024.1

# Data Collection
praw>=7.7.1
feedparser>=6.0.10
arxiv>=2.1.0
requests>=2.31.0

# Email
jinja2>=3.1.2

# AI Summarization
openai>=1.58.1

# Development
python-dotenv>=1.0.0
